refactored database lib
