+++ src/index.js
@@ -2,10 +2,10 @@ import React from "react";
 import ReactDom from "react-dom";
 import "babel-polyfill";
 import App from "./app";
-import { DatabaseProvider, initDatabase } from "./libs/db";
+import Database, { DatabaseProvider } from "./libs/database";
 
 ReactDom.render(
-  <DatabaseProvider.Provider value={initDatabase()}>
+  <DatabaseProvider.Provider value={Database()}>
     <App />
   </DatabaseProvider.Provider>,
   document.getElementById("app")
