+++ src/libs/database/add-interested-user.js
@@ -1,0 +1,14 @@
+import { INTERESTED_USERS } from "./constants";
+
+const addInterestedUser = ({ database }) => async ({ email }) => {
+  try {
+    const docRef = await database.collection(INTERESTED_USERS).add({
+      email
+    });
+    return docRef;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+export default addInterestedUser;
