+++ src/libs/database/database-provider.js
@@ -1,0 +1,4 @@
+import React from "react";
+const DatabaseProvider = React.createContext();
+
+export default DatabaseProvider;
