+++ src/libs/database/initialise-database.js
@@ -1,0 +1,13 @@
+import firebase from "firebase";
+import "firebase/firestore";
+
+const initialiseDatabase = () =>
+  firebase
+    .initializeApp({
+      apiKey: process.env.FIREBASE_API_KEY,
+      authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+      projectId: process.env.FIREBASE_PROJECT_ID
+    })
+    .firestore();
+
+export default initialiseDatabase;
